refactor: separate CategorySection from CreateCookiePage
