+++ package.json
@@ -16,8 +16,8 @@
     "axios": "^0.25.0",
     "caver-js": "^1.6.7",
     "caver-js-ext-kas": "^1.9.0",
-    "klip-sdk": "^2.0.1",
     "date-fns": "^2.28.0",
+    "klip-sdk": "^2.0.1",
     "polished": "^4.1.4",
     "qrcode.react": "^1.0.1",
     "react": "^17.0.2",
